Add files via upload
the two .spectra file is the input files, I defined a struct to record the column of the file. As there're strings in the struct, so I fail to use fwrite and fread to implement block I/O. But I tried the method of serializing them with char buffer and then use fwrite or fread to reduce I/O times.  Speed :
fgets > fscanf > ifstream > (fread + sscanf);  Speed : (sprintf+fwrite) > fprintf > ofstream.